Refactor endpoint rule compile helpers
